Some small modifications on how player templates work.
If the first mobile is to be treated as a player template, it must be initially detached from rooms since addNewPlayerASAP will give it a room and there could be duplicate issues otherwise (this does not apply to "proper" player templates defined with the specific tag, although these might be deprectated in the future. Player templates can be defined on a room I think).