Items now have a getTotalWeight() that includes the weight of their contents if they are containers. When adding/removing an Item to an Inventory, now its weight is incremented/decremented by the item's total weight instead of by the regular weight.
This will make weights work as expected in the presence of containers, as long as we don't put an item A inside B when B is itself inside some C. This could cause inconsistencies and should be avoided, at least at the moment (a more flexible weight system could be implemented in the future by not storing the weight, and calculating it on the fly instead).

Fixes issue #90.