Fixed a bug that would cause events to be processed by SwingEditBoxListener in an inconsistent state if the ENTER key was held, causing getInput to return null.
Now, the inTransitoryState flag in SwingEditBoxListener marks that a waitKeyPress() has been called but we have not yet established the wait key press state in the edit box and edit box listen.

If this flag is not used, we are at risk of the following when holding the ENTER key for a while:
1. waitKeyPress() is called, but
2. a pending ActionEvent (ENTER keypress) is processed before the change of "press any key" state is realised
3. this ActionEvent notifies the waiting waitKeyPress() method, so it returns and the game engine thread continues its run
4. then, the change of "press any key" state happens, so we are at an inconsistent state: the edit box thinks we are
waiting for a key, but the game engine has moved on and issues a getInput() call
5. the edit box's next ENTER key press notifies the getInput() call and it returns null.

Fixes Issue #144.