Lock and unlock actions have been fixed in the case where you have the key but locking/unlocking fails anyway. Fixes issue #176.
Code of open, close, lock and unlock has been modernised. Now they take the creature who opens/closes/locks/unlocks as a parameter so we can correctly use "viewer" in the associated descriptions. Also, the old composite description format has been discontinued (it was used in a single entity in Morluck's and has been removed from there, so this should not be a problem).