Fixed a memory leak caused by a combination of recent changes to the Messages class and bad threading.
Even though the Messages instance associated with a World was cleared from cache on the GameEngineThread:exitNow() method, the game engine thread keeps executing until the current update() cycle is complete. This could (and did) cause the Messages instance to be re-cached before the GameEngineThread really finished, therefore creating the memory leak since the Messages instance had a reference to the World.

To fix it, exitNow() now does a join to wait for the game engine thread to die. But this also required to modify several ColoredSwingClient methods to return immediately if the client is disabled (isDisconnected()). Without this, it could happen that the event dispatching thread was waiting (join()) for the game engine thread to die, and the game engine thread was waiting for the event dispatching thread in some invokeAndWait() - for example, when creating a frame (this is done in Morluck's intro so it would hang AGE when rebooting before the intro was finished).

Now the issue seems to be fixed although we need to be on the look for more possible deadlocks caused by the added join().